Added MIT License
Added the MIT License to the repository to specify the terms of use for the project. The MIT License allows users to freely use, modify, and distribute the software for any purpose, provided that the original copyright notice and license are included with any copies or substantial portions of the software.